requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
facenet-pytorch==2.5.3
filelock==3.19.1
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pillow==10.2.0
requests==2.32.5
scikit-learn==1.4.2
scipy==1.16.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.15.0
urllib3==2.5.0
